Add `menu_grid_type` .ini setting
This allows it to either be set to `square_icons` for the existing 8x4 grid of square icons, or `banners` for the new 3x4 grid of banners (which have an aspect ratio of 3:1).

If not specified, it’ll default to the square icons (the current behaviour).

(cherry picked from commit b2ba26bcbb23a265098ba1236e86304619ebac77)